Fix/Issue 176
